gate warp features behind build tag
